dir_py/real_estate_ai_project/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.2
aiohttp==3.10.8
aiosignal==1.3.1
alembic==1.13.3
altair==5.4.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.6.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
asyncer==0.0.7
attrs==23.2.0
auth0-python==4.7.2
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.35.34
botocore==1.35.34
bs4==0.0.2
build==1.2.2
CacheControl==0.14.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
chainlit==1.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
cleo==2.1.0
click==8.1.7
cloudpickle==2.2.1
cohere==5.11.0
colorama==0.4.6
coloredlogs==15.0.1
configparser==7.1.0
contourpy==1.2.1
crashtest==0.4.1
crewai==0.67.1
crewai-tools==0.12.1
cryptography==43.0.1
cycler==0.12.1
databricks-sdk==0.33.0
dataclasses-json==0.6.7
datasets==3.0.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.9
distlib==0.3.8
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docx2txt==0.8
duckduckgo_search==6.2.13
dulwich==0.21.7
durationpy==0.7
embedchain==0.1.122
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.112.4
fastavro==1.9.7
fastjsonschema==2.20.0
filelock==3.16.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.20.0
google-auth==2.35.0
google-cloud-aiplatform==1.69.0
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
gptcache==0.1.44
graphene==3.3
graphql-core==3.2.4
graphql-relay==3.2.0
greenlet==3.1.1
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.25.1
humanfriendly==10.0
hyperframe==6.0.1
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
iniconfig==2.0.0
installer==0.7.0
instructor==1.3.3
ipython==8.28.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpickle==3.3.0
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==24.3.1
kiwisolver==1.4.5
kubernetes==31.0.0
lancedb==0.5.7
langchain==0.2.16
langchain-cohere==0.1.9
langchain-community==0.2.17
langchain-core==0.2.41
langchain-experimental==0.0.65
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
langsmith==0.1.131
Lazify==0.4.0
linkify-it-py==2.0.3
litellm==1.48.6
literalai==0.0.607
lxml==5.3.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
marvin==2.3.7
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mem0ai==0.1.17
mlflow==2.16.2
mlflow-skinny==2.16.2
mmh3==5.0.1
mock==4.0.3
monotonic==1.6
more-itertools==10.5.0
mpire==2.10.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.17
mypy-extensions==1.0.0
narwhals==1.8.4
neo4j==5.25.0
nest-asyncio==1.6.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.50.2
openpyxl==3.1.5
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-exporter-otlp-proto-http==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.2
parameterized==0.9.0
parso==0.8.4
partialjson==0.0.8
pathlib==1.0.1
pathos==0.3.3
pexpect==4.9.0
pillow==10.4.0
pkginfo==1.11.1
platformdirs==4.3.6
pluggy==1.5.0
poetry==1.8.3
poetry-core==1.9.0
poetry-plugin-export==1.8.0
portalocker==2.10.1
posthog==3.6.6
pox==0.3.5
ppft==1.7.6.9
praw==7.8.0
prawcore==2.4.0
primp==0.6.3
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==4.25.5
psutil==6.0.0
ptyprocess==0.7.0
pulsar-client==3.5.0
pure_eval==0.2.3
py==1.11.0
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.9.0
pylance==0.9.18
pymongo==4.10.1
pyparsing==3.1.2
pypdf==4.3.1
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pyreadline3==3.5.4
pyright==1.1.383
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.4
pytube==15.0.0
pytz==2024.1
pyvis==0.3.2
pywin32==306
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qdrant-client==1.11.3
rank-bm25==0.2.2
RapidFuzz==3.10.0
ratelimiter==1.2.0.post0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
s3transfer==0.10.2
safetensors==0.4.5
sagemaker==2.232.2
sagemaker-core==1.0.10
sagemaker-mlflow==0.1.0
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.0
selenium==4.25.0
semchunk==2.2.0
semver==3.0.2
setuptools==72.2.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
smdebug-rulesconfig==1.0.1
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
starlette==0.37.2
streamlit==1.38.0
sympy==1.13.3
syncer==2.0.3
tabulate==0.9.0
tblib==3.0.0
tenacity==8.5.0
textual==0.79.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.20.0
toml==0.10.2
tomli==2.0.1
tomlkit==0.13.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
trio==0.26.2
trio-websocket==0.11.1
trove-classifiers==2024.9.12
tweepy==4.14.0
TwitterAPI==2.8.2
typer==0.12.5
types-requests==2.32.0.20240914
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
update-checker==0.18.0
uptrace==1.26.0
urllib3==2.2.3
uvicorn==0.25.0
virtualenv==20.26.6
waitress==3.0.0
watchdog==4.0.2
watchfiles==0.20.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.2
